+++ src/parse/constants.h
@@ -10,18 +10,17 @@ Returns index of the first single line comment, starting from pointer `code`.
 
 If no comment is found, the index of the last character is returned.
 */
-long long find_line_comment(const char *code) {
-	const unsigned long long code_len=strlen(code);
+unsigned long long int find_line_comment(const char *code) {
+	const size_t code_len=strlen(code);
 
-	const unsigned long long searchable_range=(code_len - LINE_COMMENT_LEN);
+	const size_t searchable_range=(code_len - LINE_COMMENT_LEN);
 	if (searchable_range<LINE_COMMENT_LEN || searchable_range>code_len) {
 		return code_len;
 	}
 
-	unsigned long long i=0;
-
+	unsigned long long int i=0;
 	for (; (i<=searchable_range); i++) {
-		if (strncmp(LINE_COMMENT, code+i, LINE_COMMENT_LEN)==0) {
+		if (strncmp(LINE_COMMENT, code + i, LINE_COMMENT_LEN)==0) {
 			break;
 		}
 	}
